Eliminate AbstractRLPInput abstraction to reduce technical debt
Merge AbstractRLPInput (607 lines) and BytesValueRLPInput (84 lines) into a
single concrete class implementing RLPInput directly. This eliminates the
unnecessary abstraction layer where the abstract class had only one concrete
implementation.

Changes:
- Merged all AbstractRLPInput logic into BytesValueRLPInput
- Made previously protected methods private
- Removed AbstractRLPInput.java file
- All public APIs remain unchanged (zero breaking changes)
- All tests pass unchanged

Benefits:
- Eliminates 15-30 minutes investigation time per RLP modification
- Reduces cognitive load when working with RLP functionality
- Simplifies maintenance with single concrete implementation
- No performance or functionality impact

Fixes #8884
